+++ packages/ingest-mongodb-public/src/sources/mongodb-dot-com/index.ts
@@ -1,0 +1,1 @@
+export * from "./WebDataSource";
